Multiply By 3 V2
