+++ src/Multiply.php
@@ -6,6 +6,13 @@ use Moontoast\Math\BigNumber;
 
 class Multiply
 {
+    public function multiplyBy3(int $number)
+    {
+        $number = new BigNumber($number);
+        $number->multiply(3);
+        return intval($number->getValue());
+    }
+
     public function multiplyBy5(int $number)
     {
         $number = new BigNumber($number);
