+++ tests/MultiplyBy3CommandTest.php
