+++ tests/MultiplyTest.php
@@ -26,11 +26,22 @@ class MultiplyTest extends TestCase
         $this->assertSame(20, $this->multiply->multiplyBy5(4));
     }
 
-    public function testC(){
+    public function testC()
+    {
 	$this->assertSame(35, $this->multiply->multiplyBy7(5));
     }
 
-    public function testD(){
+    public function testD()
+    {
 	$this->assertSame(28, $this->multiply->multiplyBy7(4));
+    }
+    public function testE()
+    {
+	$this->assertSame(15, $this->multiply->multiplyBy3(5));
+    }
+
+    public function testF()
+    {
+	$this->assertSame(24, $this->multiply->multiplyBy3(8));
     }
 }
