--- tests/composer.json
@@ -1,24 +1,0 @@
-{
-  "require": {
-    "moontoast/math": "1.1.2",
-    "symfony/console": "~2.7.0",
-    "symfony/dependency-injection": "~2.0.15"
-  },
-  "autoload": {
-    "psr-4": {
-      "App\\": "src/"
-    }
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "Tests\\": "tests/"
-    }
-  },
-  "config": {
-    "bin-dir": "bin/"
-  },
-  "require-dev": {
-    "sensiolabs/security-checker": "^5.0",
-    "phpunit/phpunit": "^6.5"
-  }
-}
